fix: use an available node type for elasticache update test
Not all node types are available to be used on elasticache,
and they need to be taken from an api. While the test was using
this workflow for creating the cache, it was not using it
for updating it.
Modify the tests to pick a valid node as well for that test.

Signed-off-by: Yolanda Robla <info@ysoft.biz>