feat: enable long queries on TheButton (#1582)
* feat: enable long queries on TheButton

When generating queries, we are passing them via GET, which
rapidly is hitting max limit for parameters. We can instead
generate a cookie with the md5 for the query and store the full
select there. Next, the frontend should be able to identify
the cookie and translate into proper query.

Signed-off-by: Yolanda Robla <yolanda@iasql.com>

* compress string to allow longer queries

* add static dependencies

Signed-off-by: Yolanda Robla <yolanda@iasql.com>